requirements.txt
tkmacosx==1.0.1
Pillow==8.2.0
atomicwrites==1.4.0
attrs==20.3.0
colorama==0.4.4
iniconfig==1.1.1
packaging==20.9
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.2
toml==0.10.2

websockets~=8.1